Fix compatibility with Python 3.15
