+++ tests/unit/conftest.py
@@ -7,67 +7,12 @@ from pathlib import Path
 import pytest
 from flexmock import flexmock
 
-import packit
-import packit.config.aliases
 from packit.config import Config
-from packit.config.aliases import Distro
 from packit.distgit import DistGit
 from packit.local_project import LocalProjectBuilder
 from packit.sync import SyncFilesItem
 from packit.upstream import GitUpstream
 from tests.spellbook import CRONIE, get_test_config, initiate_git_repo
-
-
-@pytest.fixture()
-def mock_get_aliases():
-    mock_aliases_module = flexmock(packit.config.aliases)
-    mock_aliases_module.should_receive("get_aliases").and_return(
-        {
-            "fedora-all": [
-                Distro("fedora-29", "f29"),
-                Distro("fedora-30", "f30"),
-                Distro("fedora-31", "f31"),
-                Distro("fedora-32", "f32"),
-                Distro("fedora-33", "f33"),
-                Distro("fedora-rawhide", "rawhide"),
-            ],
-            "fedora-stable": [Distro("fedora-31", "f31"), Distro("fedora-32", "f32")],
-            "fedora-branched": [Distro("fedora-33", "f33")],
-            "fedora-development": [
-                Distro("fedora-33", "f33"),
-                Distro("fedora-rawhide", "rawhide"),
-            ],
-            "epel-all": [
-                Distro("epel-6", "el6"),
-                Distro("epel-7", "epel7"),
-                Distro("epel-8", "epel8"),
-                Distro("epel-9", "epel9"),
-                Distro("epel-10.0", "epel10.0"),
-                Distro("epel-10.1", "epel10"),
-            ],
-            "epel-10-all": [
-                Distro("epel-10.0", "epel10.0"),
-                Distro("epel-10.1", "epel10"),
-            ],
-            "epel-10-branched": [
-                Distro("epel-10.0", "epel10.0"),
-            ],
-            "epel-10": [
-                Distro("epel-10.1", "epel10"),
-            ],
-            "opensuse-leap-all": [
-                "opensuse-leap-15.5",
-                "opensuse-leap-15.4",
-                "opensuse-leap-15.3",
-            ],
-            "opensuse-all": [
-                "opensuse-tumbleweed",
-                "opensuse-leap-15.5",
-                "opensuse-leap-15.4",
-                "opensuse-leap-15.3",
-            ],
-        },
-    )
 
 
 @pytest.fixture
